# Optimized the calculation of quantile hazard maps