There is a 7x speedup for EUR:
```
# before
| ncalls      | cumtime | path                                        |
|-------------+---------+---------------------------------------------|
| 4614        | 188.384 | stats.py:177(quantile_curve)                |
| 1661040     | 121.965 | stats.py:206(<listcomp>)                    |
| 3326131     | 17.056  | fromnumeric.py:51(_wrapfunc)                |
| 1661065     | 11.506  | fromnumeric.py:1012(argsort)                |
| 1538        | 6.843   | getters.py:220(get_hcurve)                  |

# after
| ncalls      | cumtime | path                                        |
|-------------+---------+---------------------------------------------|
| 4614        | 27.579  | stats.py:177(quantile_curve)                |
| 3326131     | 15.649  | fromnumeric.py:51(_wrapfunc)                |
| 1661065     | 10.250  | fromnumeric.py:1012(argsort)                |
| 1538        | 6.613   | getters.py:220(get_hcurve)                  |
```